add player attack.. enemies now take damage and get destroyed once below 0 health. add wall of corruption.

+++ Phoenix-of-the-Forest/Assets/Enemy_Data/Scripts/Enemy.cs
@@ -13,7 +13,7 @@ public enum EnemyType {
 }
 
 public abstract class Enemy : ScriptableObject {
-   public float maxHealth;
+   public int maxHealth;
    public int damage;
    public float moveSpeed;
    public float attackRange;
